🔀 :: 리프레시 토큰 만료시간 수정

+++ src/main/kotlin/com/project/goms/domain/auth/entity/RefreshToken.kt
@@ -14,5 +14,5 @@ data class RefreshToken(
     val accountIdx: UUID,
 
     @TimeToLive(unit = TimeUnit.SECONDS)
-    val expiredAt: Int
+    val expiredAt: Long
 )
